# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql"
)

type ApiToken struct {
	ID          string        `json:"id"`
	UserID      int64         `json:"user_id"`
	Name        string        `json:"name"`
	TokenPrefix string        `json:"token_prefix"`
	TokenHash   string        `json:"token_hash"`
	Scope       string        `json:"scope"`
	LastUsedAt  sql.NullInt64 `json:"last_used_at"`
	ExpiresAt   sql.NullInt64 `json:"expires_at"`
	CreatedAt   int64         `json:"created_at"`
	RevokedAt   sql.NullInt64 `json:"revoked_at"`
}

type AuditLog struct {
	ID         int64          `json:"id"`
	UserID     sql.NullInt64  `json:"user_id"`
	Action     string         `json:"action"`
	EntityType string         `json:"entity_type"`
	EntityID   sql.NullInt64  `json:"entity_id"`
	Details    sql.NullString `json:"details"`
	CreatedAt  int64          `json:"created_at"`
}

type Deployment struct {
	ID            int64          `json:"id"`
	ReleaseID     int64          `json:"release_id"`
	EnvironmentID int64          `json:"environment_id"`
	Status        string         `json:"status"`
	StartedAt     sql.NullInt64  `json:"started_at"`
	FinishedAt    sql.NullInt64  `json:"finished_at"`
	CreatedAt     int64          `json:"created_at"`
	Forced        int64          `json:"forced"`
	Note          sql.NullString `json:"note"`
}

type DeploymentApproval struct {
	ID                   int64         `json:"id"`
	DeploymentID         int64         `json:"deployment_id"`
	ApprovedBy           string        `json:"approved_by"`
	ApprovedAt           int64         `json:"approved_at"`
	ApproverUserID       sql.NullInt64 `json:"approver_user_id"`
	RequiredApproverRole string        `json:"required_approver_role"`
}

type DeploymentLog struct {
	ID           int64          `json:"id"`
	DeploymentID int64          `json:"deployment_id"`
	StepName     sql.NullString `json:"step_name"`
	Line         string         `json:"line"`
	CreatedAt    int64          `json:"created_at"`
}

type Environment struct {
	ID          int64          `json:"id"`
	Name        string         `json:"name"`
	Description sql.NullString `json:"description"`
	Tags        sql.NullString `json:"tags"`
	CreatedAt   int64          `json:"created_at"`
}

type GlobalNotification struct {
	ID                int64          `json:"id"`
	SlackWebhookUrl   sql.NullString `json:"slack_webhook_url"`
	NotifyEmails      sql.NullString `json:"notify_emails"`
	GotifyUrl         sql.NullString `json:"gotify_url"`
	GotifyToken       sql.NullString `json:"gotify_token"`
	DiscordWebhookUrl sql.NullString `json:"discord_webhook_url"`
	UpdatedAt         int64          `json:"updated_at"`
}

type Lifecycle struct {
	ID          int64          `json:"id"`
	Name        string         `json:"name"`
	Description sql.NullString `json:"description"`
	CreatedAt   int64          `json:"created_at"`
}

type LifecycleStage struct {
	ID               int64 `json:"id"`
	LifecycleID      int64 `json:"lifecycle_id"`
	EnvironmentID    int64 `json:"environment_id"`
	SortOrder        int64 `json:"sort_order"`
	RequiresApproval int64 `json:"requires_approval"`
}

type NotificationEvent struct {
	ID            int64         `json:"id"`
	EventType     string        `json:"event_type"`
	DeploymentID  sql.NullInt64 `json:"deployment_id"`
	ProjectID     sql.NullInt64 `json:"project_id"`
	EnvironmentID sql.NullInt64 `json:"environment_id"`
	Message       string        `json:"message"`
	Results       string        `json:"results"`
	CreatedAt     int64         `json:"created_at"`
}

type Project struct {
	ID                int64          `json:"id"`
	Name              string         `json:"name"`
	Description       sql.NullString `json:"description"`
	CreatedAt         int64          `json:"created_at"`
	LifecycleID       sql.NullInt64  `json:"lifecycle_id"`
	SlackWebhookUrl   sql.NullString `json:"slack_webhook_url"`
	NotifyEmails      sql.NullString `json:"notify_emails"`
	GotifyUrl         sql.NullString `json:"gotify_url"`
	GotifyToken       sql.NullString `json:"gotify_token"`
	DiscordWebhookUrl sql.NullString `json:"discord_webhook_url"`
}

type ProjectMember struct {
	ProjectID int64  `json:"project_id"`
	UserID    int64  `json:"user_id"`
	Role      string `json:"role"`
	CreatedAt int64  `json:"created_at"`
}

type Release struct {
	ID        int64  `json:"id"`
	ProjectID int64  `json:"project_id"`
	Version   string `json:"version"`
	StepsJson string `json:"steps_json"`
	CreatedAt int64  `json:"created_at"`
}

type ReleaseVariable struct {
	ID            int64          `json:"id"`
	ReleaseID     int64          `json:"release_id"`
	Name          string         `json:"name"`
	Value         sql.NullString `json:"value"`
	EnvironmentID sql.NullInt64  `json:"environment_id"`
	CreatedAt     int64          `json:"created_at"`
	Secret        int64          `json:"secret"`
}

type ScheduledDeployment struct {
	ID            int64          `json:"id"`
	ProjectID     int64          `json:"project_id"`
	ReleaseID     int64          `json:"release_id"`
	EnvironmentID int64          `json:"environment_id"`
	Cron          string         `json:"cron"`
	NextRunAt     int64          `json:"next_run_at"`
	Enabled       int64          `json:"enabled"`
	LastFiredAt   sql.NullInt64  `json:"last_fired_at"`
	Note          sql.NullString `json:"note"`
	CreatedAt     int64          `json:"created_at"`
	UpdatedAt     int64          `json:"updated_at"`
}

type Session struct {
	ID        string         `json:"id"`
	UserID    int64          `json:"user_id"`
	CsrfToken string         `json:"csrf_token"`
	CreatedAt int64          `json:"created_at"`
	ExpiresAt int64          `json:"expires_at"`
	IpAddress sql.NullString `json:"ip_address"`
	UserAgent sql.NullString `json:"user_agent"`
}

type Step struct {
	ID             int64  `json:"id"`
	ProjectID      int64  `json:"project_id"`
	Name           string `json:"name"`
	ScriptBody     string `json:"script_body"`
	SortOrder      int64  `json:"sort_order"`
	CreatedAt      int64  `json:"created_at"`
	TimeoutSeconds int64  `json:"timeout_seconds"`
	MaxRetries     int64  `json:"max_retries"`
}

type StepTemplate struct {
	ID         int64  `json:"id"`
	Name       string `json:"name"`
	ScriptBody string `json:"script_body"`
	CreatedAt  int64  `json:"created_at"`
}

type StepTemplateVersion struct {
	ID            int64  `json:"id"`
	TemplateID    int64  `json:"template_id"`
	VersionNumber int64  `json:"version_number"`
	Name          string `json:"name"`
	ScriptBody    string `json:"script_body"`
	CreatedAt     int64  `json:"created_at"`
}

type User struct {
	ID           int64         `json:"id"`
	Email        string        `json:"email"`
	PasswordHash string        `json:"password_hash"`
	Name         string        `json:"name"`
	Role         string        `json:"role"`
	CreatedAt    int64         `json:"created_at"`
	UpdatedAt    int64         `json:"updated_at"`
	LastLoginAt  sql.NullInt64 `json:"last_login_at"`
}

type Variable struct {
	ID            int64          `json:"id"`
	ProjectID     int64          `json:"project_id"`
	Name          string         `json:"name"`
	Value         sql.NullString `json:"value"`
	EnvironmentID sql.NullInt64  `json:"environment_id"`
	CreatedAt     int64          `json:"created_at"`
	Secret        int64          `json:"secret"`
}