# internal/swagger/spec.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "durpdeploy REST API.",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
...
    "/projects/{id}/templates-picker": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "templates"
        ],
        "summary": "List step templates for the picker.",
        "operationId": "templatesPicker",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "StepTemplateListResponse is a list of step templates.",
            "schema": {
              "$ref": "#/definitions/StepTemplateListResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/templates": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "templates"
        ],
        "summary": "List all step templates.",
        "operationId": "listTemplates",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "StepTemplateListResponse is a list of step templates.",
            "schema": {
              "$ref": "#/definitions/StepTemplateListResponse"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "templates"
        ],
        "summary": "Create a step template.",
        "operationId": "createTemplate",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StepTemplateRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "StepTemplate is a reusable step template.",
            "schema": {
              "$ref": "#/definitions/StepTemplate"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "409": {
            "description": "Conflict error response.",
            "schema": {
              "$ref": "#/definitions/ConflictError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/templates/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "templates"
        ],
        "summary": "Get a step template.",
        "operationId": "getTemplate",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "StepTemplate is a reusable step template.",
            "schema": {
              "$ref": "#/definitions/StepTemplate"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "templates"
        ],
        "summary": "Update a step template.",
        "operationId": "updateTemplate",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StepTemplateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "StepTemplate is a reusable step template.",
            "schema": {
              "$ref": "#/definitions/StepTemplate"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "409": {
            "description": "Conflict error response.",
            "schema": {
              "$ref": "#/definitions/ConflictError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "templates"
        ],
        "summary": "Delete a step template.",
        "operationId": "deleteTemplate",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "EmptyResponse is returned for no-content endpoints.",
            "schema": {
              "$ref": "#/definitions/EmptyResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/templates/{id}/history": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "templates"
        ],
        "summary": "List version history for a step template.",
        "operationId": "getTemplateHistory",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "StepTemplateVersionListResponse is a list of step template versions.",
            "schema": {
              "$ref": "#/definitions/StepTemplateVersionListResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "BadRequestError": {
      "type": "object",
      "title": "Bad request error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerBadRequestError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
    "ConflictError": {
      "type": "object",
      "title": "Conflict error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerConflictError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "EmptyResponse": {
      "type": "object",
      "title": "EmptyResponse is returned for no-content endpoints.",
      "x-go-name": "swaggerEmptyResponse",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "NotFoundError": {
      "type": "object",
      "title": "Not found error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerNotFoundError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "ServerError": {
      "type": "object",
      "title": "Server error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerServerError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "StepTemplate": {
      "type": "object",
      "title": "StepTemplate is a reusable step template.",
      "properties": {
        "created_at": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreatedAt"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "script_body": {
          "type": "string",
          "x-go-name": "ScriptBody"
        }
      },
      "x-go-name": "swaggerStepTemplate",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
    "StepTemplateListResponse": {
      "type": "array",
      "title": "StepTemplateListResponse is a list of step templates.",
      "items": {
        "$ref": "#/definitions/StepTemplate"
      },
      "x-go-name": "swaggerStepTemplateListResponse",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
    "StepTemplateRequest": {
      "type": "object",
      "title": "StepTemplateRequest is the body for create/update step template.",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "script_body": {
          "type": "string",
          "x-go-name": "ScriptBody"
        }
      },
      "x-go-name": "swaggerStepTemplateRequest",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
    "StepTemplateVersion": {
      "type": "object",
      "title": "StepTemplateVersion is a historical version of a step template.",
      "properties": {
        "created_at": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreatedAt"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "script_body": {
          "type": "string",
          "x-go-name": "ScriptBody"
        },
        "template_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "TemplateID"
        },
        "version_number": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "VersionNumber"
        }
      },
      "x-go-name": "swaggerStepTemplateVersion",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
    "StepTemplateVersionListResponse": {
      "type": "array",
      "title": "StepTemplateVersionListResponse is a list of step template versions.",
      "items": {
        "$ref": "#/definitions/StepTemplateVersion"
      },
      "x-go-name": "swaggerStepTemplateVersionListResponse",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "UnauthorizedError": {
      "type": "object",
      "title": "Unauthorized error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerUnauthorizedError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}